Update from original master

+++ RAW_Sensor.cs
@@ -1,14 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
-using System.Text;
 using System.Windows.Forms;
 using ZedGraph;
-using AGaugeApp;
-using System.IO.Ports;
-using System.Threading;
 using MissionPlanner.Utilities;
 
 namespace MissionPlanner
@@ -275,9 +268,9 @@ namespace MissionPlanner
 
         private void CMB_rawupdaterate_SelectedIndexChanged(object sender, EventArgs e)
         {
-            MainV2.comPort.MAV.cs.ratesensors = (byte) int.Parse(CMB_rawupdaterate.Text);
+            MainV2.comPort.MAV.cs.ratesensors = int.Parse(CMB_rawupdaterate.Text);
             MainV2.comPort.requestDatastream(MAVLink.MAV_DATA_STREAM.RAW_SENSORS,
-                (byte) int.Parse(CMB_rawupdaterate.Text)); // request raw sensor
+                int.Parse(CMB_rawupdaterate.Text)); // request raw sensor
         }
 
         System.IO.StreamWriter sw = null;
